requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
google-generativeai==0.7.2
google-auth==2.29.0
google-auth-httplib2==0.2.0
requests==2.32.3
xhtml2pdf==0.2.11
Pillow==10.4.0
whitenoise==6.7.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
dj-database-url==2.2.0
